refactor: rename module name - virtual-thread -> spring-virtual-thread
