+++ spring-java/spring-virtual-thread/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'spring-virtual-thread'
