+++ spring-java/spring-virtual-thread/src/main/java/com/example/spring_virtual_thread/SpringVirtualThreadApplication.java
@@ -1,13 +1,13 @@
-package com.example.virtual_thread;
+package com.example.spring_virtual_thread;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class VirtualThreadApplication {
+public class SpringVirtualThreadApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(VirtualThreadApplication.class, args);
+		SpringApplication.run(SpringVirtualThreadApplication.class, args);
 	}
 
 }
