+++ spring-java/spring-virtual-thread/src/main/java/com/example/spring_virtual_thread/controller/request/ExecutionEmptyRequest.java
@@ -1,4 +1,4 @@
-package com.example.virtual_thread.controller.request;
+package com.example.spring_virtual_thread.controller.request;
 
 import lombok.Data;
