+++ spring-java/spring-virtual-thread/src/test/java/com/example/spring_virtual_thread/SpringVirtualThreadApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.virtual_thread;
+package com.example.spring_virtual_thread;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class VirtualThreadApplicationTests {
+class SpringVirtualThreadApplicationTests {
 
 	@Test
 	void contextLoads() {
