--- spring-java/virtual-thread/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'virtual-thread'
